+++ PropertyInspector/CHANGELOG.md
@@ -4,11 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0]
 ### Added
- - 
+ - `InspectorNameAttribute` alias.
+ - `HeaderAttribute` alias.
+ - Support for both `HeaderAttribute` & `InspectorNameAttribute` aliases.
+ - Added `tooltip` to `HeaderAttribute` alias and support for it in the property drawer.
+ - Dependencies to package.json.
+ - Added complete replacement versions of all currently aliased attributes. These don't rely on inheritance, and thus are more stable. Currently unsupported.
 ### Changed
- - 
+ - All aliases now derived from the type they are aliases of.
+ - Changed how `GetPropertyHeight` determines property height.
+### Fixed
+ - `SpaceAttribute` support: multiple instances on 1 member were ignored, now they aren't.
+ - `SpaceAttribute` and `HeaderAttribute` now respect `PropertyAttribute.order`.
+ - `EditorGUIUtility.standardVerticalSpacing` now taken into account for spacing and property height.
+ - Property drawer breaking upon property order change.
+### Removed
+ - Removed `IHasPropertyInspectors` and `PropertyInspectorDrawer` - so much has changed, it's best to redo them entirely later than keep them as-is.
 
 ## [0.1.0]
 ### Added